requirements.txt
pillow
matplotlib
scikit-image
numpy
h5py
tqdm
tensorflow==1.9.0
protobuf==3.6.0
keras==2.2.5
pyqt5
opencv-python
imgaug